perf: Fill a nested described member in place, build a tree's objects in linear time
Two things two outside benchmarks found:

- read_into() had in-place branches for containers and optionals, but a described
  or converted type fell through to as<T>() - built afresh, filled, moved in - so
  every nested member's vectors and strings were reallocated on every read where
  the top level already filled in place. It calls serializer<T>::read now, and a
  member read again keeps the capacity it has.
- value::object looked for the name before every insertion, so a tree read from
  a document cost N-squared comparisons for an object of N members; an object of
  17,850 took 220 ms. Building appends and coalesces duplicates once at the end,
  in linear time above sixteen members, keeping the last value under a repeated
  name where the first was: 4.8 ms.

And two spellings other libraries have that this one lacked: decode_into(text,
value) and write(value, string&), for many documents through one object or one
buffer. A float member is written with a float's shortest digits, taking an
exponent from ten million as Żmij does for a float, where it was widened to a
double and written at seventeen; the two writers agree on every float tried.

+++ include/serpent/json/text.hpp
@@ -89,7 +89,7 @@ template<std::floating_point T>
         std::memcpy(to, "null", 4);
         return to + 4;
     }
-    return to + serpent::detail::write_real(to, static_cast<double>(value));
+    return to + serpent::detail::write_real(to, value);
 }
 
 } // namespace serpent::json
